Added: requirements.txt. Added: A build file to automate production builds. Changed: Made app.py compatible with production versions. Added: Release version of .envs. Fixed: A crash when deleting the main group. Fixed: Build issues with the frontend. Added: Icon and name for frontend build.

+++ backend/controllers/app/app_controller.py
@@ -10,6 +10,9 @@ class AppController(BaseController):
 
         @self._app.route(f"{base_route}/quit", methods=["POST"])
         def quit():
+            if os.getenv("BUILD_MODE") == "DEBUG":
+                return
+
             pid = os.getpid()
             os.kill(pid, signal.SIGINT)
             return make_response("", 200)
